Changing test case status without editing test case

+++ src/hooks/useTestCases.ts
@@ -1,12 +1,24 @@
 import { useCallback } from "react";
 import { supabase } from "../supabaseClient";
 
-export function useTestCases(checkedTestCases: number[]) {
+export function useTestCases(checkedTestCases: number | number[] | undefined) {
+  const idsAsArray =
+    checkedTestCases === undefined
+      ? []
+      : Array.isArray(checkedTestCases)
+        ? checkedTestCases
+        : [checkedTestCases];
+
   const deleteTestCases = useCallback(async () => {
+    if (idsAsArray.length === 0) {
+      console.error("Empty array");
+      return;
+    }
+
     const { data, error } = await supabase
       .from("test_cases")
       .delete()
-      .in("id", checkedTestCases)
+      .in("id", idsAsArray)
       .select();
 
     if (error) console.error(error);
@@ -15,10 +27,15 @@ export function useTestCases(checkedTestCases: number[]) {
 
   const updateTestCasesStatus = useCallback(
     async (newStatus: "To Do" | "Passed" | "Failed" | "Skipped") => {
+      if (idsAsArray.length === 0) {
+        console.error("Empty array");
+        return;
+      }
+
       const { data, error } = await supabase
         .from("test_cases")
         .update({ status: newStatus })
-        .in("id", checkedTestCases)
+        .in("id", idsAsArray)
         .select();
 
       if (error) console.error(error);
